add Kopano to the list of organizations
Hi,
at Kopano we are also true believers of open source and therefore have all our software projects available with an agpl license. While we do sell a subscription for our software (which gives access to qa'ed software packages and package repositories as well as support) you can download nightly releases and run it with the same feature set as our customers do.