Add Redash
Hi,

I'm the creator of Redash, the open source project and the founder of Redash, the company.

Redash runs in a similar model to Sentry -- open source SaaS, with 100% of the code in the open source version (except for things like the billing code).

I'm happy to answer questions or clarify if anything is unclear. 

Thanks.